Removed static container, need to decide what to do with status controller, this is used when installed locally, but now the solution is moving towards a central server and using the network shares to access each test station.
The central solution allows the user to easily see the status of each station and run tests on each one. The working process is currently check the home page to see what stations are free and run the test case on 1 or more stations, once the case is complete an email containing a summary and a link to a teamCity run, from team city detailed test data can be reviewed.